requirements.txt
fastapi[standard]
python-dotenv
asyncpg
sqlalchemy
pydantic-settings
greenlet
orjson
pytest
testcontainers[postgresql]
pytest-asyncio
slowapi